pyproject.toml
[tool.poetry]
name = "text-to-audio"
version = "0.1.0"
description = "A python script to convert text to audio"
authors = ["MCDAngelo"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pypdf = "^4.3.1"
ipython = "^8.26.0"
google-cloud-texttospeech = "^2.16.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
